requirements.txt
greenlet==2.0.2
pan-python==0.22.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
xmltodict==0.13.0
